pyproject.toml
[tool.poetry]
name = "foundational"
version = "0.1.0"
description = ""
authors = ["rockey20230213 <rocket23>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pipecat-ai = {extras = ["daily"], version = "^0.0.49"}
python-dotenv = "^1.0.1"
onnxruntime = "^1.20.1"
openai = "^1.55.2"
websockets = "13.1"
cartesia = "^1.2.0"
livekit = "0.17.5"
tenacity = "^9.0.0"
livekit-api = "0.7.1"
aiohttp = "3.10.10"
loguru = "^0.7.2"
fal-client = "^0.5.6"
faster-whisper = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
